refactor: interfaces
