+++ rslang/src/pages/Sprint/Sprint.interface.ts
@@ -1,0 +1,12 @@
+export interface IStatistics {
+  id: string;
+  word: string;
+  audio: string;
+  transcription: string;
+  wordTranslate: string;
+  result: boolean;
+}
+
+export interface ISprint {
+  isGameOpenFromMenu: boolean;
+}
