+++ rslang/src/pages/Sprint/components/buttons/buttonSpeak/ButtonSpeak.interface.ts
@@ -1,0 +1,5 @@
+import { IWordsResponse } from '../../../../../features/words/wordsSlice.interface';
+
+export interface IButtonSpeak {
+  audio: IWordsResponse | undefined | string;
+}
