+++ rslang/src/pages/Sprint/components/greetings/Greetings.interface.ts
@@ -1,0 +1,6 @@
+import { MouseEvent } from 'react';
+
+export interface IGreetings {
+  onClick: (event: MouseEvent<HTMLButtonElement>) => void;
+  isGameOpenFromMenu: boolean;
+}
