+++ rslang/src/pages/Sprint/components/statistics/Statistics.interface.ts
@@ -1,0 +1,12 @@
+export interface IStatistics {
+  statistics: {
+    id: string;
+    word: string;
+    audio: string;
+    transcription: string;
+    wordTranslate: string;
+    result: boolean;
+  }[];
+  endGame: () => void;
+  resetGame: () => void;
+}
